% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply Filter to Image}
\usage{
apply_filter(bitmap, filter_func, ...)
}
\arguments{
\item{bitmap}{A bitmap array from magick.}

\item{filter_func}{The filter function to apply.}

\item{...}{Additional arguments for the filter function.}
}
\value{
A magick image object.
}
\description{
Apply Filter to Image
}
