% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutoff_filter}
\alias{cutoff_filter}
\title{Cutoff Filter}
\usage{
cutoff_filter(rgb_vec, cutoff = 127)
}
\arguments{
\item{rgb_vec}{A vector of RGB values.}

\item{cutoff}{A numeric cutoff between 0 and 255.}
}
\value{
A vector of raw values representing a color (black or magenta).
}
\description{
Cutoff Filter
}
