% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{greyscale_filter}
\alias{greyscale_filter}
\title{Grayscale Filter}
\usage{
greyscale_filter(rgb_vec)
}
\arguments{
\item{rgb_vec}{A vector of RGB values (as raw or integer).}
}
\value{
A vector with grayscale RGB values.
}
\description{
Grayscale Filter
}
